AI/MLAgent/requirements.txt
mlagents==1.1.0
torch-geometric
numpy
matplotlib 
tensorboard 
pandas
opencv-python
pyzmq
ultralytics
-r torch-requirement.txt